requirements.txt
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.0
pytest-cov==4.1.0